# openapi/atlas.v1.yaml
openapi: 3.1.0
info:
  title: Atlas DAM API
  version: "1.0"
  description: >
    API-first DAM (doc 22). Source de vérité unique : UI (WASM), SDK, CLI et
    serveur MCP sont générés/dérivés de ce contrat. Servi localement (/openapi.json),
    zéro dépendance externe.
  license:
    name: Apache-2.0
servers:
  - url: /v1
...
paths:
...
  /search:
    post:
      tags: [search]
      operationId: search
      summary: Recherche hybride conversationnelle (doc 04/25)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SearchRequest" }
      responses:
        "200":
          description: Résultats
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SearchResponse" }
        "422": { $ref: "#/components/responses/Problem" }
...
  /search/suggest:
    get:
      tags: [search]
      operationId: suggest
      summary: Autocomplétion par préfixe sur les titres d'assets (doc 25 §5)
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
          description: "Préfixe à compléter (vide → aucune suggestion)"
        - name: limit
          in: query
          schema: { type: integer, default: 8, maximum: 20 }
      responses:
        "200":
          description: Suggestions
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Suggestions" }
        "500": { $ref: "#/components/responses/Problem" }
...
components:
  responses:
    Problem:
      description: Erreur (RFC 9457)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    RightsStatus:
      type: string
      enum: [valid, expiring, expired, none]
...
    StructuredFilter:
      type: object
      description: "Filtres structurés (déduits ou explicites, doc 25 §4.1)"
      properties:
        has_people: { type: boolean, nullable: true }
        orientation: { type: string, nullable: true }
        rights_status: { $ref: "#/components/schemas/RightsStatus" }
        type: { type: array, items: { type: string } }
    SearchRequest:
      type: object
      required: [query]
      properties:
        query: { type: string }
        mode: { type: string, enum: [natural, lexical, example], default: natural }
        example_asset_id:
          type: string
          format: uuid
          nullable: true
          description: "Asset source pour mode:example (doc 25 §4.2)"
        filters:
          allOf: [ { $ref: "#/components/schemas/StructuredFilter" } ]
          nullable: true
          description: "Filtres explicites ; priment sur les filtres déduits (interpreted_query éditable)"
        page_size: { type: integer, default: 50, maximum: 200 }
        cursor: { type: string, nullable: true }
    SearchResultItem:
      type: object
      properties:
        asset_id: { type: string, format: uuid }
        score: { type: number }
        title: { type: string }
        rights_status: { $ref: "#/components/schemas/RightsStatus" }
    InterpretedQuery:
      type: object
      description: "Requête interprétée, renvoyée au client et éditable (doc 25 §4.1)"
      properties:
        semantic_text: { type: string }
        filters: { $ref: "#/components/schemas/StructuredFilter" }
        confidence: { type: number }
        editable: { type: boolean }
    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items: { $ref: "#/components/schemas/SearchResultItem" }
        interpreted_query: { $ref: "#/components/schemas/InterpretedQuery" }
        facets:
          type: object
          description: "Agrégations par facette sur l'ensemble autorisé (doc 25 §4.5) ; omis si vide"
          additionalProperties:
            type: array
            items: { $ref: "#/components/schemas/FacetCount" }
        next_cursor: { type: string, nullable: true }
        query_hash:
          type: string
          description: "Hash de la requête (doc 25 §4.4/§6) ; à renvoyer sur POST /v1/search/click"
        degraded: { type: boolean }
...
    FacetCount:
      type: object
      properties:
        value: { type: string }
        count: { type: integer, format: int64, minimum: 0 }
      required: [value, count]
    Suggestions:
      type: object
      properties:
        suggestions:
          type: array
          items: { type: string }
      required: [suggestions]
...
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        trace_id: { type: string }